refactor: Separate FMDN code into dedicated fmdn/ module
Reorganize FMDN (Find My Device Network) functionality into a clean,
self-contained module to improve code organization and maintainability
for upstream PR.

Changes:
- Created fmdn/ submodule with clear separation of concerns:
  - extraction.py: EID extraction logic (moved from fmdn.py)
  - manager.py: EID resolution tracking (moved from bermuda_fmdn_manager.py)
  - integration.py: FMDN-coordinator integration layer (NEW)
  - __init__.py: Public API exports

- Extracted 10 FMDN methods from coordinator.py into FmdnIntegration class:
  - get_resolver(), format_metadevice_address()
  - normalize_eid_bytes(), extract_eids()
  - process_resolution(), process_resolution_with_status()
  - register_source(), handle_advertisement()
  - prune_source(), discover_metadevices()

- Updated coordinator.py:
  - Replaced self.fmdn_manager with self.fmdn (FmdnIntegration)
  - Removed EidResolver Protocol (now in integration.py)
  - Removed ~260 lines of FMDN code

- Updated diagnostics.py to use coordinator.fmdn.manager

Benefits:
- FMDN code isolated in single module (~850 lines)
- Reduced coordinator.py complexity
- Cleaner separation for potential future extraction
- Easier testing and maintenance
